// src/app/core/api/model/catalogDependencyDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CatalogDependencyDto { 
    ref: string;
    as: string;
    required: boolean;
    reuseExisting: boolean;
}