// src/app/core/api/model/dependencyChoiceDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DependencyChoiceDto { 
    /**
     * Matches spec.dependencies[].as in the manifest
     */
    alias: string;
    mode: DependencyChoiceDto.ModeEnum;
    /**
     * Required when mode=REUSE_EXISTING. ApplicationEntity id of the existing building block to reuse.
     */
    existingApplicationId?: string;
}
export namespace DependencyChoiceDto {
    export const ModeEnum = {
        Dedicated: 'DEDICATED',
        ReuseExisting: 'REUSE_EXISTING'
    } as const;
    export type ModeEnum = typeof ModeEnum[keyof typeof ModeEnum];
}